requirements.txt
# AI Operations Assistant - E-Commerce Inventory Chatbot
# LangChain + ChromaDB + OpenAI GPT-4o-mini + Voice + ETL Automation

# Core LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# Vector Database
chromadb>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# OpenAI (includes Whisper & TTS)
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Rich terminal output
rich>=13.0.0

# Audio Recording & Playback
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1

# Web Scraping
selenium>=4.0.0
webdriver-manager>=4.0.0

# ETL Automation & Scheduling
apscheduler>=3.10.0
pyyaml>=6.0

# File Watching
watchdog>=3.0.0

# Web UI (Flask + WebSocket)
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Testing
pytest>=8.0.0

